backend/requirements.txt
Django==3.2.16
djangorestframework==3.12.4
django-cors-headers==3.13.0
psycopg2-binary==2.9.3 
django-filter==23.4
djoser==2.1.0
django-colorfield==0.11.0
gunicorn==20.1.0
python-dotenv==0.20.0
Pillow==10.1.0